requirements.txt
streamlit==1.35.0
pdfplumber
pdfminer.six
python-docx==1.1.0
pyspellchecker==0.8.3
langdetect==1.0.9
spacy==3.8.7
numpy>=1.26,<2
pandas==2.2.3
pyinstaller==6.14.1
